adding append previous

+++ tensordb/file_cache_tensor_client.py
@@ -35,7 +35,7 @@ class FileCacheTensorClient:
     Parameters
     ----------
 
-    synchronizer_mode: Literal["manual", "automatic"] = "automatic"
+    synchronizer_mode: Literal["delayed", "automatic"] = "automatic"
         If this option is enable then all the writes on the base_map are also executed on the remote_map
 
     tensor_lock: ContextManager = None
@@ -48,93 +48,185 @@ class FileCacheTensorClient:
             remote_client: TensorClient,
             local_client: TensorClient,
             tensor_lock: ContextManager,
-            synchronizer_mode: Literal["manual", "automatic"] = "automatic",
+            checksum_path: str,
+            synchronizer_mode: Literal["delayed", "automatic"] = "automatic",
     ):
         self.remote_client = remote_client
         self.local_client = local_client
         self.synchronizer_mode = synchronizer_mode
         self.tensor_lock = tensor_lock
+        self.checksum_map = self.local_client.base_map.sub_map(checksum_path)
 
-    def synchronize_tensor(
+    def merge(
             self,
             path: str,
-            mode: Literal["fetch", "merge"],
             force: bool = False,
-            only_definition: bool = False
+            only_definition: bool = False,
     ):
-        if not force and self.synchronizer_mode == "manual" and mode == "merge":
+        if not force and self.synchronizer_mode == "delayed":
             return
 
-        local_client, remote_client = self.local_client, self.remote_client
-        if remote_client.exist(path):
-            if mode == "merge":
-                remote_client, local_client = self.local_client, self.remote_client
+        if not self.local_client.exist(path):
+            raise ValueError(f"The path {path} does not exist in local so it is impossible to merge it")
 
-            local_definition = local_client.get_tensor_definition(path)
-            remote_definition = remote_client.get_tensor_definition(path)
+        if not self.remote_client.exist(path):
+            force = True
+
+        local_definition = self.local_client.get_tensor_definition(path)
+
+        if "modification_date" not in local_definition.metadata:
+            raise ValueError(f"There is no modification date on the local definition metadata")
+
+        if not force:
             local_modification_date = pd.Timestamp(local_definition.metadata["modification_date"])
+            remote_definition = self.remote_client.get_tensor_definition(path)
             remote_modification_date = pd.Timestamp(remote_definition.metadata["modification_date"])
 
             if local_modification_date == remote_modification_date:
                 return
 
-            if remote_modification_date > local_modification_date and mode == "merge":
-                raise ValueError(f"During the merge process the remote tensor was modified, "
-                                 f"this can be produced by an incorrect behaviour of the tensor_lock")
+            if local_modification_date < remote_modification_date:
+                raise ValueError(f"The remote tensor was updated last time on {remote_modification_date} "
+                                 f"while the local copy was updated on {local_modification_date}, please update "
+                                 f"first the local copy before merging it.")
 
-            if local_modification_date > remote_modification_date and self.synchronizer_mode == "manual":
-                return
-
-            local_client.upsert_tensor(remote_definition)
+        self.remote_client.upsert_tensor(local_definition)
 
         if not only_definition:
-            remote_map = remote_client.base_map.sub_map(path)
-            local_map = local_client.base_map.sub_map(path)
-            remote_map.copy_to_mapping(local_map, remote_modification_date)
+            Mapping.synchronize(
+                remote_map=self.remote_client.base_map.sub_map(path),
+                local_map=self.local_client.base_map.sub_map(path),
+                checksum_map=self.checksum_map.sub_map(path),
+                force=force,
+                to_local=False
+            )
+
+    def fetch(
+            self,
+            path: str,
+            force: bool = False,
+            only_definition: bool = False,
+    ):
+        if not self.remote_client.exist(path):
+            return
+
+        remote_definition = self.remote_client.get_tensor_definition(path)
+        if not self.local_client.exist(path):
+            force = True
+
+        if "modification_date" not in remote_definition.metadata:
+            force = True
+            self.remote_client.update_tensor_metadata(path, {"modification_date": str(pd.Timestamp.now())})
+            remote_definition = self.remote_client.get_tensor_definition(path)
+
+        if not force:
+            remote_modification_date = pd.Timestamp(remote_definition.metadata["modification_date"])
+            local_definition = self.local_client.get_tensor_definition(path)
+            local_modification_date = pd.Timestamp(local_definition.metadata["modification_date"])
+
+            if local_modification_date == remote_modification_date:
+                return
+
+            if local_modification_date > remote_modification_date:
+                if self.synchronizer_mode == "delayed":
+                    return
+
+                raise ValueError(f"The remote tensor was updated last time on {remote_modification_date} "
+                                 f"while the local copy was updated on {local_modification_date}, this means "
+                                 f"that they are out of sync and this is not possible "
+                                 f"if the synchronizer mode is not delayed")
+
+        self.local_client.upsert_tensor(remote_definition)
+        if not only_definition:
+            Mapping.synchronize(
+                remote_map=self.remote_client.base_map.sub_map(path),
+                local_map=self.local_client.base_map.sub_map(path),
+                checksum_map=self.checksum_map.sub_map(path),
+                force=force,
+                to_local=True
+            )
+
+    def _exec_callable(
+            self,
+            func: str,
+            path: str,
+            fetch: bool,
+            merge: bool,
+            only_read: bool,
+            force: bool = False,
+            drop_checksums: bool = False,
+            **kwargs
+
+    ):
+        func = getattr(self.local_client, func)
+        with self.tensor_lock(path):
+            if not only_read:
+                self.update_tensor_metadata(path, {"modification_date": str(pd.Timestamp.now())})
+            if fetch:
+                self.fetch(path, force=force)
+            if drop_checksums:
+                self.checksum_map.rmdir(path)
+            result = func(path=path, **kwargs)
+            if merge:
+                self.merge(path, force=force)
+            return result
 
     def read(
             self,
             path: Union[str, TensorDefinition, xr.DataArray, xr.Dataset],
             **kwargs
     ) -> Union[xr.DataArray, xr.Dataset]:
-        with self.tensor_lock(path):
-            self.synchronize_tensor(path, "fetch")
-            return self.local_client.read(path=path, **kwargs)
+        return self._exec_callable(
+            "read",
+            path=path,
+            fetch=True,
+            merge=False,
+            only_read=True,
+            **kwargs
+        )
 
     def store(
             self,
             path: Union[str, TensorDefinition, xr.DataArray, xr.Dataset],
             **kwargs
     ) -> Union[xr.DataArray, xr.Dataset]:
-        with self.tensor_lock(path):
-            self.update_tensor_metadata(path, {"modified_date": str(pd.Timestamp.now())})
-            delayed = self.local_client.store(path=path, **kwargs)
-            self.synchronize_tensor(path, "merge")
-            return delayed
+        return self._exec_callable(
+            "store",
+            path=path,
+            fetch=True,
+            merge=True,
+            only_read=False,
+            drop_checksums=True,
+            **kwargs
+        )
 
     def update(
             self,
             path: Union[str, TensorDefinition, xr.DataArray, xr.Dataset],
             **kwargs
     ) -> Union[xr.DataArray, xr.Dataset]:
-        with self.tensor_lock(path):
-            self.synchronize_tensor(path, "fetch")
-            self.update_tensor_metadata(path, {"modified_date": str(pd.Timestamp.now())})
-            delayed = self.local_client.update(path=path, **kwargs)
-            self.synchronize_tensor(path, "merge")
-            return delayed
+        return self._exec_callable(
+            "update",
+            path=path,
+            fetch=True,
+            merge=True,
+            only_read=False,
+            **kwargs
+        )
 
     def upsert(
             self,
             path: Union[str, TensorDefinition, xr.DataArray, xr.Dataset],
             **kwargs
     ) -> Union[xr.DataArray, xr.Dataset]:
-        with self.tensor_lock(path):
-            self.synchronize_tensor(path, "fetch")
-            self.update_tensor_metadata(path, {"modified_date": str(pd.Timestamp.now())})
-            delayed = self.local_client.upsert(path=path, **kwargs)
-            self.synchronize_tensor(path, "merge")
-            return delayed
+        return self._exec_callable(
+            "upsert",
+            path=path,
+            fetch=True,
+            merge=True,
+            only_read=False,
+            **kwargs
+        )
 
     def exist(
             self,
@@ -149,32 +241,29 @@ class FileCacheTensorClient:
             only_local: bool = True,
             **kwargs
     ) -> List[AbstractWritableDataStore]:
-        with self.tensor_lock(path):
-            if not only_local:
-                self.get_client(True).drop(path=path, **kwargs)
-            return self.get_client(False).drop(path=path, **kwargs)
+        return self._exec_callable(
+            "drop",
+            path=path,
+            fetch=True,
+            merge=True,
+            only_read=False,
+            **kwargs
+        )
 
     @validate_arguments
     def create_tensor(self, definition: TensorDefinition):
-        delayed = self.local_client.create_tensor(definition=definition)
-        self.synchronize_tensor(path=path, mode="merge", force=True)
-        return delayed
+        with self.tensor_lock(path):
+            if path in self.remote_client:
+                raise KeyError(f"Overwrite the tensor definition on the path {path} "
+                               f"can produce synchronization problems")
+            self.remote_client.create_tensor(definition)
+            self.local_client.create_tensor(definition)
 
     @validate_arguments
     def get_tensor_definition(self, path: str) -> TensorDefinition:
         with self.tensor_lock(path):
-            self.synchronize_tensor(path, "fetch", only_definition=True)
+            self.fetch(path, force=False)
             return self.local_client.get_tensor_definition(path)
-
-
-    @validate_arguments
-    def update_tensor_metadata(self, path: str, new_metadata: Dict[str, Any]):
-        with self.tensor_lock(path):
-            self.synchronize_tensor(path, "fetch", only_definition=True)
-            return self.local_client.update_tensor_metadata(path=path)
-
-    def get_all_tensors_definition(self) -> List[TensorDefinition]:
-        return self.remote_client.get_all_tensors_definition()
 
     @validate_arguments
     def delete_tensor(
@@ -184,9 +273,10 @@ class FileCacheTensorClient:
             only_local: bool = True,
     ) -> Any:
         with self.tensor_lock(path):
+            if self.local_client.exist(path):
+                self.local_client.delete_tensor(path=path, only_data=only_data)
             if not only_local:
-                self.remote_client.delete_tensor(path, only_data=only_data)
-            return self.local_client.delete_tensor(path=path, only_data=only_data)
+                self.remote_client.delete_tensor(path=path, only_data=only_data)
 
     @validate_arguments
     def delete_tensors(
@@ -195,17 +285,22 @@ class FileCacheTensorClient:
             only_data: bool = False,
             only_local: bool = True
     ):
-        with self.tensor_lock(path):
-            if not only_local:
-                self.remote_client.delete_tensors(paths, only_data=only_data)
-            return self.local_client.delete_tensors(paths, only_data=only_data)
+        for path in paths:
+            try:
+                self.delete_tensor(path, only_data=only_data, only_local=only_local)
+            except KeyError:
+                pass
 
     @validate_arguments
     def get_storage(
             self,
             path: Union[str, TensorDefinition]
     ) -> BaseStorage:
-        with self.tensor_lock(path):
-            self.synchronize_tensor(path, "fetch", only_definition=True)
-            return self.local_client.get_storage(path=path)
+        return self._exec_callable(
+            "get_storage",
+            path=path,
+            fetch=True,
+            merge=False,
+            only_read=True,
+        )
